Refactor : client/config에 input길이 관련 상수들 client/inputConstrains로 이동
이유 : input 제약 사항과 관련된 constants를 관리하기 위해
