+++ client/src/constants/inputConstraints.js
@@ -1,0 +1,5 @@
+const DEFAULT_TEXT_INPUT_MAX_LENGTH = 100;
+const NICKNAME_LENGTH = 8;
+const MAX_CHAT_LENGTH = 40;
+
+export { DEFAULT_TEXT_INPUT_MAX_LENGTH, NICKNAME_LENGTH, MAX_CHAT_LENGTH };
