+++ client/src/presentation/components/MessageInput.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
 import Input from '@material-ui/core/Input';
 import { STYLE_COLORS } from '../../utils';
-import { DEFAULT_TEXT_INPUT_MAX_LENGTH } from '../../config';
+import { DEFAULT_TEXT_INPUT_MAX_LENGTH } from '../../constants/inputConstraints';
 
 const useStyles = makeStyles({
   input: {
