+++ client/src/utils/browserLocalStorage.js
@@ -1,4 +1,5 @@
-import { NICKNAME_LENGTH, LOCALSTORAGE_NICKNAME_KEY } from '../config';
+import { LOCALSTORAGE_NICKNAME_KEY } from '../config';
+import { NICKNAME_LENGTH } from '../constants/inputConstraints';
 
 const getNickname = () => {
   if (!localStorage) return '';
